dbConnector generic class
