+++ packages/external-db-mysql/lib/index.js
@@ -6,28 +6,17 @@ const init = require('./connection_provider')
 const DatabaseOperations = require('./mysql_operations')
 const { supportedOperations } = require('./supported_operations')
 const { MySqlConfigValidator } = require ('./config_validator')
+const { DbConnector } = require('velo-external-db-commons')
 
 const driver = () => require('../tests/drivers/sql_filter_transformer_test_support')
 const opsDriver = () => require('../tests/drivers/db_operations_test_support')
 
 
-class MySqlConnector {
+class MySqlConnector extends DbConnector {
     constructor() {
+        super(MySqlConfigValidator, init)
         this.type = 'mysql'
-        this.initialized = false
     }
-    
-    async initialize(config, options) {
-        const { dataProvider, schemaProvider, databaseOperations, connection, cleanup } = await init(config, options)
-        this.dataProvider = dataProvider
-        this.schemaProvider = schemaProvider
-        this.databaseOperations = databaseOperations
-        this.connection = connection
-        this.cleanup = cleanup
-        this.configValidator = new MySqlConfigValidator(config)  
-        this.initialized = true
-        return { dataProvider, schemaProvider, databaseOperations, connection, cleanup }
-    }    
 }
 
 const mySqlFactory = async(config, options) => {
