+++ packages/velo-external-db-commons/lib/index.js
@@ -2,10 +2,12 @@ const errors = require('./errors')
 const schemaCommons = require('./schema_commons')
 const dataCommons = require('./data_commons')
 const configCommons = require ('./config_commons')
+const DbConnector = require ('./db_connector')
 
 module.exports = {
     errors,
     ...schemaCommons,
     ...dataCommons,
-    ...configCommons
+    ...configCommons,
+    DbConnector
 }
